refactor: remove obsolate structs
